+++ packages/frontend/src/components/Dialog/CloseButton.tsx
@@ -3,11 +3,19 @@ import React from 'react'
 import HeaderButton from './HeaderButton'
 
 import type { ButtonHTMLAttributes } from 'react'
+import useTranslationFunction from '../../hooks/useTranslationFunction'
 
 export default function CloseButton(
   props: ButtonHTMLAttributes<HTMLButtonElement>
 ) {
+  const tx = useTranslationFunction()
+
   return (
-    <HeaderButton aria-label='Close' icon='cross' iconSize={26} {...props} />
+    <HeaderButton
+      aria-label={tx('close')}
+      icon='cross'
+      iconSize={26}
+      {...props}
+    />
   )
 }