+++ packages/frontend/src/components/Dialog/FooterActionButton.tsx
@@ -2,29 +2,16 @@ import React from 'react'
 
 import Button from '../Button'
 
-import type { MouseEventHandler, PropsWithChildren } from 'react'
+import type { ButtonProps } from '../Button'
 
 import styles from './styles.module.scss'
 
-type Props = PropsWithChildren<{
-  onClick: MouseEventHandler<HTMLButtonElement>
-  disabled?: boolean
-  danger?: boolean
-}>
-
 export default function FooterActionButton({
   children,
-  onClick,
-  disabled = false,
-  danger = false,
-}: Props) {
+  ...props
+}: ButtonProps) {
   return (
-    <Button
-      className={styles.footerActionButton}
-      onClick={onClick}
-      type={danger ? 'danger' : undefined}
-      disabled={disabled}
-    >
+    <Button className={styles.footerActionButton} {...props}>
       {children}
     </Button>
   )