+++ packages/frontend/src/components/LoginForm.tsx
@@ -2,7 +2,6 @@
 
 import { C, DcEventType } from '@deltachat/jsonrpc-client'
 import React, { useEffect, useRef, useState } from 'react'
-import { Collapse } from '@blueprintjs/core'
 import { useDebouncedCallback } from 'use-debounce/lib'
 
 import {
@@ -10,7 +9,6 @@ import {
   DeltaPasswordInput,
   DeltaSelect,
   DeltaProgressBar,
-  DeltaSwitch,
 } from './Login-Styles'
 import { ClickableLink } from './helpers/ClickableLink'
 import { Credentials } from '../types-app'
@@ -24,11 +22,13 @@ import Dialog, {
   FooterActionButton,
   FooterActions,
 } from './Dialog'
+import Collapse from './Collapse'
 import { I18nContext } from '../contexts/I18nContext'
 import useTranslationFunction from '../hooks/useTranslationFunction'
 import { getDeviceChatId, saveLastChatId } from '../backend/chat'
 
 import type { DialogProps } from '../contexts/DialogContext'
+import SettingsSwitch from './Settings/SettingsSwitch'
 
 const log = getLogger('renderer/loginForm')
 
@@ -47,11 +47,8 @@ export function defaultCredentials(credentials?: Credentials): Credentials {
     send_port: '',
     send_security: '',
     smtp_certificate_checks: '',
-    socks5_enabled: '0',
-    socks5_host: '',
-    socks5_port: '',
-    socks5_user: '',
-    socks5_password: '',
+    proxy_enabled: '0',
+    proxy_url: '',
   }
   return { ...defaultCredentials, ...credentials }
 }
@@ -144,11 +141,8 @@ export default function LoginForm({ credentials, setCredentials }: LoginProps) {
     send_server,
     send_port,
     send_security,
-    socks5_enabled,
-    socks5_host,
-    socks5_port,
-    socks5_user,
-    socks5_password,
+    proxy_enabled,
+    proxy_url,
   } = credentials
 
   // We assume that smtp_certificate_checks has the same value.
@@ -191,15 +185,16 @@ export default function LoginForm({ credentials, setCredentials }: LoginProps) {
           )}
 
           <p className='text'>{tx('login_no_servers_hint')}</p>
-          <div
+          <button
             className='advanced'
+            aria-controls='advanced-collapse'
             onClick={() => setUiShowAdvanced(!uiShowAdvanced)}
             id={'show-advanced-button'}
           >
             <div className={`advanced-icon ${uiShowAdvanced && 'opened'}`} />
             <p>{tx('menu_advanced')}</p>
-          </div>
-          <Collapse isOpen={uiShowAdvanced}>
+          </button>
+          <Collapse id='advanced-collapse' isOpen={uiShowAdvanced}>
             <br />
             <p className='delta-headline'>{tx('login_inbox')}</p>
 
@@ -307,50 +302,26 @@ export default function LoginForm({ credentials, setCredentials }: LoginProps) {
                 {tx('accept_invalid_certificates')}
               </option>
             </DeltaSelect>
-            <DeltaSwitch
-              id='socks5_enabled'
-              label={tx('login_socks5_use_socks5')}
-              value={socks5_enabled}
-              onChange={isTrue =>
-                _handleCredentialsChange('socks5_enabled', isTrue ? '1' : '0')
+            <SettingsSwitch
+              label={tx('proxy_use_proxy')}
+              value={proxy_enabled === '1'}
+              onChange={(isTrue: boolean) =>
+                _handleCredentialsChange('proxy_enabled', isTrue ? '1' : '0')
               }
             />
-            {socks5_enabled === '1' && (
+            {proxy_enabled === '1' && (
               <>
                 <p className='text'>
-                  {tx('login_socks5_experimental_warning')}
+                  Proxy support is currently experimental. Please use at your
+                  own risk. If you type in an address in the e-mail field, there
+                  will be a DNS lookup that won't get tunneled through proxy.
                 </p>
+                <p className='text'>{tx('proxy_add_explain')}</p>
                 <DeltaInput
-                  key='socks5_host'
-                  id='socks5_host'
-                  placeholder={tx('default_value', 'localhost')}
-                  label={tx('login_socks5_host')}
-                  value={socks5_host}
-                  onChange={handleCredentialsChange}
-                />
-                <DeltaInput
-                  key='socks5_port'
-                  id='socks5_port'
-                  placeholder={tx('default_value', '9150')}
-                  label={tx('login_socks5_port')}
-                  type='number'
-                  min='0'
-                  max='65535'
-                  value={socks5_port}
-                  onChange={handleCredentialsChange}
-                />
-                <DeltaInput
-                  key='socks5_user'
-                  id='socks5_user'
-                  label={tx('login_socks5_user')}
-                  value={socks5_user}
-                  onChange={handleCredentialsChange}
-                />
-                <DeltaInput
-                  key='socks5_password'
-                  id='socks5_password'
-                  label={tx('login_socks5_password')}
-                  value={socks5_password}
+                  key='proxy_url'
+                  id='proxy_url'
+                  label={tx('proxy_add_url_hint')}
+                  value={proxy_url}
                   onChange={handleCredentialsChange}
                 />
               </>
@@ -472,7 +443,7 @@ export function ConfigureProgressDialog({
       </DialogBody>
       <DialogFooter>
         <FooterActions>
-          <FooterActionButton danger onClick={onCancel}>
+          <FooterActionButton styling='danger' onClick={onCancel}>
             {tx('cancel')}
           </FooterActionButton>
         </FooterActions>