+++ packages/frontend/src/components/Settings/Appearance.tsx
@@ -1,6 +1,5 @@
 import { join } from 'path'
 import React, { useEffect, useState } from 'react'
-import { Icon } from '@blueprintjs/core'
 
 import { ThemeManager } from '../../ThemeManager'
 import { runtime } from '@deltachat-desktop/runtime-interface'
@@ -21,6 +20,7 @@ import { DialogContent } from '../Dialog'
 import useTranslationFunction from '../../hooks/useTranslationFunction'
 import useDialog from '../../hooks/dialog/useDialog'
 import { LastUsedSlot, rememberLastUsedPath } from '../../utils/lastUsedPaths'
+import Icon from '../Icon'
 import Callout from '../Callout'
 
 const log = getLogger('renderer/settings/appearance')
@@ -329,16 +329,22 @@ function SettingsTrayIcon({
 }) {
   const tx = useTranslationFunction()
 
+  const isBrowser = runtime.getRuntimeInfo().target === 'browser'
+  const disabled = settingsStore.rc.minimized || isBrowser
+
   return (
     <>
       <DesktopSettingsSwitch
         settingsKey='minimizeToTray'
         label={tx('pref_show_tray_icon')}
-        disabled={settingsStore.rc.minimized}
-        disabledValue={settingsStore.rc.minimized}
+        disabled={disabled}
+        disabledValue={settingsStore.rc.minimized && !isBrowser}
       />
       {settingsStore.rc.minimized && (
         <Callout>{tx('explain_desktop_minimized_disabled_tray_pref')}</Callout>
+      )}
+      {isBrowser && (
+        <Callout>{tx('explain_desktop_browser_disabled_tray_pref')}</Callout>
       )}
     </>
   )